requirements.txt
flask
transformers
pytorch torchvision torchaudio cudatoolkit=11.3